chore: 상위 폴더 builder, viewer 생성 (#44)
