+++ src/components/wishpool/builder/create/step/Step1.tsx
@@ -1,21 +1,28 @@
 import ButtonContainer from '@/components/wishpool/form/create/ButtonContainer';
 import InputBox from '@/components/wishpool/form/create/InputBox';
 import Question from '@/components/wishpool/form/create/Question';
-import { StepProps } from '@/types/wishpool/form/stepProps';
+import { StepProps } from '@/types/wishpool/builder/stepProps';
 
-const Step4 = ({ onPrev, onNext }: StepProps) => {
+const Step1 = ({ onPrev, onNext }: StepProps) => {
   return (
     <section className="px-[2rem]">
-      <div className="mt-[1.6rem]">
+      <div className="mt-[2.8rem]">
         <Question
           required={true}
-          question="참여자의 선물 리스트를 언제까지 받을까요?"
+          question="위시풀을 받을 생일자의 정보를 알려 주세요."
         />
       </div>
 
+      <div className="mt-[4rem]">
+        <InputBox
+          content="생일자 이름"
+          placeholder="생일자의 이름을 알려 주세요."
+          maxLength={20}
+        />
+      </div>
       <div className="mt-[3.2rem]">
         <InputBox
-          content="선물 리스트 마감일"
+          content="곧 다가올 생일"
           calendar={true}
           placeholder="YY/MM/DD"
         />
@@ -29,4 +36,4 @@ const Step4 = ({ onPrev, onNext }: StepProps) => {
   );
 };
 
-export default Step4;
+export default Step1;
