+++ src/components/wishpool/builder/create/step/Step2.tsx
@@ -1,7 +1,7 @@
 import ButtonContainer from '@/components/wishpool/form/create/ButtonContainer';
 import InputBox from '@/components/wishpool/form/create/InputBox';
 import Question from '@/components/wishpool/form/create/Question';
-import { StepProps } from '@/types/wishpool/form/stepProps';
+import { StepProps } from '@/types/wishpool/builder/stepProps';
 
 const Step2 = ({ onPrev, onNext }: StepProps) => {
   return (
