+++ src/components/wishpool/builder/create/step/Step3.tsx
@@ -1,7 +1,7 @@
 import Icon from '@/components/common/Icon';
 import ButtonContainer from '@/components/wishpool/form/create/ButtonContainer';
 import Question from '@/components/wishpool/form/create/Question';
-import { StepProps } from '@/types/wishpool/form/stepProps';
+import { StepProps } from '@/types/wishpool/builder/stepProps';
 
 const Step3 = ({ onPrev, onNext }: StepProps) => {
   return (
