+++ src/components/wishpool/builder/intro/InfoNotice.tsx
