+++ src/constants/wishpool/builder/infoNoticeList.tsx
